enable_lbaas flag in the testbed.py to enable/disable lbaas feature.
We used haproxy flag to decide on LBaaS feature provisioing, which is misleading
and also it configures haproxy on openstack/config and vrouter nodes. To avoid this
added env.enable_lbaas varaible which defaults to Fasle is added, Also used this valiable
to decide on installign the additional package in the vrouter node that are requird for
LBaaS feature.
Adding the new section SERVICE-INSTALNCE in the contrail-vrouter-agent.conf, after upgrade.

Change-Id: I7c081c627c97d5ab229a75fc11b7b9aa4c278185