Use control_data ip to make remote_syslog work in multiinterface setup
In fabfile/tasks/provision.py, we populate at line 1304 control_data interface ip and
try to access the connect_map_dict with management ip,Which is in a multi interface setup.
Fixed it by using control_ip and  pointing to the correct script location.

Closes-Bug: 1383441

enable_lbaas flag in the testbed.py to enable/disable lbaas feature.

We used haproxy flag to decide on LBaaS feature provisioing, which is misleading
and also it configures haproxy on openstack/config and vrouter nodes. To avoid this
added env.enable_lbaas varaible which defaults to Fasle is added, Also used this valiable
to decide on installign the additional package in the vrouter node that are requird for
LBaaS feature.
Adding the new section SERVICE-INSTALNCE in the contrail-vrouter-agent.conf, after upgrade.

Change-Id: I066f2c258e53b347d3147bfafc8bc80876a9787c