Storage Chassis support
Closes-Bug: #1385504
Feature: Need functionality to provision storage nodes with the notion
of chassis in the crushmap. This is required for systems like Quanta,
where a single chassis has 4 nodes built into it. The replica for an
object should not reside on the same chassis.
Support: Added support for the above. Changes done to modify the crush
map and have chassis configuration in addition to the host
configuration. The Chassis configuration will have the list of hosts
that share the common chassis. The leaf node of pool rules are
configured to use Chassis instead of hosts. This way the replica will be
put under a different chassis.
Also provided a command in fab 'fab storage_chassis_configure' to
dynamically add the support in an existing cluster without destroying.
Test: Verified the provisioning with a 2Node chassis configuration with
and without HDD/SSD pools. Verified the pg map and confirmed that the
replica is not placed under the same chassis.

Change-Id: I094c862c8849fe7de6f5bd0756b03d08898845d5