Fixes: Openstack HA - API operations become really slow and sluggish
Excerpt from HAProxy config
If the system supports it, it can be useful on big sites to raise this limit
very high so that haproxy manages connection queues, instead of leaving the
clients with unanswered connection attempts. This value should not exceed the
global maxconn.

The maxconn for MySQL is 10000 but the default global maxconn value is 1999.
In certain loaded systems we see that the reach the haproxy maxconn frontend
connection limit and hence HAProxy does not accept new connections. Increasing
the global limit to at least what the indivual servers have.

Change-Id: Ic87a1f913ab2f6ca5154a23186468bd70d4fe01e
ClosesBug: 1424153