Allow root/sudo users to execute fab commands without specifing the password in the testbed.py, Instead by specifying the ssh public key as the env.key_filename variable in testbed.py
It will be users responsibility to add the pbublic key to authorized_keys
in all the nodes in the cluster prior to provisioning using fab commands.

For sudo users all fab command can take -I arguemnt to prompt for initial sudo
password once before getting in the task, or can set NOPASSWD in the sudoers
file for the username.

Change-Id: I06522e65a0acde4c9bd955379afa0f19509a959d