Adding a task to check and report reimage status.
Adding task all_sm_reimage_status to check and report the reimage
status of the nodes. By default this runs 180 iterations at an
intervall of 10 seconds each to check if all the nodes reimaged
properly. Else it reports the status where reimage got stuck. As
debug reporting this is usefull.
Change-Id: Ic5e2dfb3bfb6065484654f340ee6e49dc747bd4e