Adding ping routine before ssh trial.
Adding a routine to check ICMP connectivity to the hosts when
fab task - wait_till_all_up is run. ping will be run to the
hosts to check connectivity before we run ssh connectivity tests.
This will just provide an additional layer of debugability to
the existing task. If ping to any of the host fails, i.e. 100%
packet loss then we exit the task.

Adding a sleep of 30 sec between reimage command issue for multi
nodes. This is a temp fix till bug 1442296 is addressed.

Change-Id: I13ea7788f4a8a24a5eeca9e7665d4b85f78f3b5c