Allow agent_name to be configured in testbed file and populate that value in tor-agent config file.
A new field named 'tor_agent_name' can now be configured under env.tor_agent stanza of testbed file. This value will be populated for agent_name field of contrail-tor-agent config file.
Also ensured that ssl configuration files are generated only once when standby configuration is specified for TOR. Before generating the files we check if any standby is configured and if so we try to copy files from standby (if they are already generated on standby).

Change-Id: If69150cf83ae158639125d4cd121ebe6e2eef34b
Closes-Bug: #1455619