Closes-Bug: 1463380 Initial checkin of cluster based reimage.
Reimage servers with cluster id if that is available as an
env variable in the testbed.py file as:
env.cluster_id='xxxx_xxxx'
If not then we follow the earlier way of node by node reimage.
In-case the testbed has a esxi server we follow node by node
reimage.

Introducing a flag of check_status=False. Enabling this would
call the all_sm_reimage_status function after reimage to track
the status of the servers through reimage_completed state.

Change-Id: Ib5c21c476f2ade37afdd3e3366443a82c4f11db3