Fixes: upgrade_contrail failed in a HA setup
In 2.2, we are re-creating the cmon_schema. The Galera cluster
was forming and hence when we try to setup cmon_schema it fails.

Before setting up cmon schema, make sure that mysql galera status
is up and running, and if its in joining state - wait for it to
completely sync.

Change-Id: Ie890f624e0402991ea9adde25a8faff33eb301d8
ClosesBug: 1462209