vcenter: Changes in vmx file of ContrailVM vmdk
Repeated characters when typing in remote console of ContrailVM.

If you are using a wide-area or low-bandwidth connection, the time delay over the
network may be long enough to cause the virtual machine to start auto-repeat.

Add below to the ContrailVm vmx file.

keyboard.typematicMinDelay = "2000000"

The delay is specified in micro-seconds, increases the repeat time to 2 seconds.

Closes-Bug: #1526186

Change-Id: I45c0012f77f5d718765cf8f65e4a343283ead171