Fixes:Openstack HA : router_id in keepalived configuration to be configurable Desscription Description: It is a configurable parameter now. ClosesBug:1379559
Note: There are totally three files as part of this change. But, for some reason these changes weren't pushed into the main line after the approval on Jul 30 2015.
Later on Feb 4 2016, it was repushed as it was a missing commit. It was resulting in a conflict because of the below reason.
Changes from Change-Id: Idaa2951f861dc5246b3b02081d1f47383694c71d against bug 1524919 brought in fabfile/utils/host.py changes resulting in conflict because of partial commit here.
So pushing only the rest of the changes that are needed. Note, these changes were already reviewed.

Change-Id: I9f91629161a05d35e7301f166d2205472f119588