Fix detach_vrouter_node task
When compute node is provisioned, the setup-vnc-compute tool of the
contrail-provisioning repo is used. It then, internally, executes
/opt/contrail/utils/provision_vrouter.py and passes as the --host-name
parameter the return value of socket.gethostname().

According to the python's documentation, socket.gethostname() may return a
fully qualified domain name or not, depending on the existing configuration.

The same applies to the socket.gethostyaddr() used here in the
detach_vrouter_node task.

However, it had always been stripped down to only the base hostname, which may
lead to discrepancy in the global config between what was set up during the
provisioning and what is looked up when detaching the node, like in the
example below:

cfgm_common.exceptions.NoIdError: Unknown id: Error: oper 1 url /fqname-to-id
body {"fq_name": ["default-global-system-config", "l4"], "type":
"virtual-router"} response Name ['default-global-system-config', 'l4'] not
found

where there is an entry for the 'l4.full.domain.com' in the
default-glogal-system-config, but not for the 'l4'.

Change-Id: I0905042c9351a0c71df47f3704b70483fe304f01
Closes-Bug: #1543542