Update HAproxy config for tor-agent to not use leastconn.
As leastconn is taking precedence and not moving the connection to
another tor-agent when the failure happens on the node with lower
number of connections, removing this from the haproxy config. The
default is round-robin. While this may lead to connections not being
balanced, it keeps the connections going in all cases.

Change-Id: I2eaee83ba532c1be837a4f402bf276c37f009d3d
partial-bug: #1545956