support to configure the same mac to all the sr-iov interfaces in vmware
The only supported bond mode in sr-iov interfaces in vmware is
active-backup. since the fail_over_mac is set as active, the bond would
use the active interface's mac to receive the packets. sicne the vrf is
with old mac, the packet would be dropped in the interface. By having
same mac on all the interfaces, it can be avoided

Change-Id: I060c1aaa7cf9765d7139b0538c18d28ab42cfa5c
Closes-bug: 1592914