vCenter-as-compute: Fix add_esxi
"fab add_vrouter_node" is used to setup a ContrailVM while adding an
ESXi to an existing cluster. ContrailVM inside the ESXi is a vrouter-only node,
manage_nova_compute and confgure_nova need to be turned off while setting it up.

This was missing in install_vrouter_node and setup_vrouter_node, added it.

Change-Id: I5ec03d2fad1309f26b5bd6af2f199dd6e1f01534
Closes-Bug: #1630001