Common package for contrail and storage
Closes-Bug: #1648703
Issue:
In existing infra, storage module is built as separate package and will
not be available as part of contrail-install packages. It requires
following additional steps during the package installation.

1. Extract the contrail-storage package
2. Will create separate repo for stroage using setup_stroage.sh

Fix:
storage module also will be built along with the contrail-install. So
above two steps will not be required any more.
In main-line, this change will be added for mitaka only. Whenever
newton is ready, the  same will be added for newton as well.
The common code changes are applicable for other openstack releases
as well. So the changes are added to accomodate the mulitple packages
for older releases (< mitaka) and single packages for newer releases
(>= mitaka)
Change-Id: If6c8207d1b76c39f0f1930706cbf9252b15835ee