Fix verify_database on centos7 and redhat
On centos7 and redhat, contrail-database is still a SYSV service and
hence its status needs to be checked by looking at the return code of
service contrail-database status rather than the output since the output
does not contain the word running. Pipe-ing head -n 10 changes the return
code to always return 0. The above was broken due to commit for bug #1638670
commit f378fac60aa4372bc50c8f96eefef1a5df0c5d3b which made setup_all work
on xenial. Revert the change to verify_service made in that commit.

Change-Id: If38f781713145d28f8725634685f8f98661bc056
Closes-Bug: #1657934