Add depends package path as arg to install_ns_agilio_nic fab task
Closes-Bug: #1665740

This changeset provides tasking that allows the cfgm node to copy
the ns-agilio-vrouter-depends-packages package to the compute nodes
and registers the package with the local apt-get repo.  Existing
fab tasking will then apt-get install the ns-agilio-vrouter-depends-packages
when needed.

The implementation was modeled from the install_pkg_all task.

Execution example
fab install_ns_agilio_nic:/tmp/ns-agilio-vrouter-depends-packages_3.1.0.0-124_amd64.deb

This changeset also enables the automated install path for
Netronome Agilio SmartNICs that are not acceleration enabled
in the testbed.  This was added to enable using the SmartNIC
as a NIC and to allow auto install of drivers and firmware.

Change-Id: Iccb9ce2ba4fd72b2e50762c74a55022adb03b5cb