Fab changes to support the scenario where esxi already added in vcenter by name
The esxi was pre-existing as attached in the snapshot by name esxi-60-172-25-47-80.
1.We searched the host in vcenter by IP as name
2.It did not get it
3.Tried to add the host to cluster
4.Failed as host with different name but same ip already present in vcenter.Vcenter threw the exception.

To make this scenario work, made the below changes
1.Added this in testbed.py

esxi_hosts = {
       'esxi-60-172-25-47-80': {
             'ip': '172.25.47.80',
             'name': 'esxi-60-172-25-47-80’,(optional)
We need to document this that we should not use this field except brownfield

We first search with by name if provided, then return the host if already presnt.
If hostname not present, use ip address

Change-Id: Ic999349bf71ff4260685a555038cdb895214fc3d
Closes-Bug: 1716167