Too many Redis connections left idle and aged
Currently options to kill the idle connections was not there.
Added a timeout of 12 hr i.e if for 12 hr a connection is idle
it will close that connection.
And setting the tcp-keepalive value as well so that if peers are
dead for that many seconds the connection will be closed.
setting this value to 300 as redis defined this in 3.2 onwards releases.

DocImpact
closes-jira-bug: JCB-218835
Implements: blueprint libvirt-xml-cpu-model

Change-Id: I45237b84d09a414f3d006b7dd776b6a1c3a8901c